// frontend/src/services/household_debt_api/api.ts
export interface CountyDebt {
	/**
	 * County
	 */
	county: string
	/**
	 * Average debt ratio
	 */
	average_debt: number
	/**
	 * FIPS code for the County
	 */
	fips_code: string
}
export interface CountyExtended {
	/**
	 * County Name
	 */
	name: string
	/**
	 * County Fips Code
	 */
	fips_code: string
	/**
	 * State Code for County
	 */
	state_code: string
}
// ...
export interface State {
	/**
	 * State Name
	 */
	name: string
	/**
	 * State Code
	 */
	code: string
	/**
	 * State Fips Code
	 */
	fips_code: string
}
export interface StateDebt {
	/**
	 * State
	 */
	state: string
	/**
	 * Average debt ratio
	 */
	average_debt: number
	/**
	 * Number of Counties
	 */
	number_of_counties: number
	/**
	 * FIPS code for the State
	 */
	fips_code: string
}